Cleanup: Last minute tweaks...
 * change a few more functions from long to size_t
 * add a few ASSERT statements where appropriate.  These make sure that
	casting a size_t to an int or long will not truncate large values.
 * refactor a few places where the size_t changes could be better
 * fix one usage of _tsplitpath_s() that was wrong.

+++ Externals/crystaledit/editlib/ccrystaleditview.cpp
@@ -2798,7 +2798,7 @@ OnUpdateToolsCharCoding (CCmdUI * pCmdUI)
   pCmdUI->Enable (IsSelection ());
 }
 
-long str_pos (LPCTSTR whole, LPCTSTR piece);
+size_t str_pos (LPCTSTR whole, LPCTSTR piece);
 
 void CCrystalEditView::
 OnToolsCharCoding ()
